Fix wrong positions of certain boundary walls
Because the judgment of `isWall` was wrong, the position of the wall
was off by one. The boundary wall was rendered at the farther position,
so it didn't connect to its neighboring walls and was seemed smaller on
the screen. Thus, I revised the `if` condition to make its position
correct.