Port the raycaster to the ARM baremetal platform
This experimental update confirms that the raycaster does not rely on
any floating-point operations and also demonstrates that it has the
potential to run in a baremetal machine, which means that the execution
environment lacks any kind of graphical user interface support and
built-in I/O operation procedures.

I wrote a custom bootcode in assembly, a framebuffer driver, a UART
driver, and an interface to the Raspberry Pi firmware's mailbox system.
I also updated the Makefile so that the user can run the baremetal
raycaster directly with the "make run" command and control the view by
sending W/A/S/D via UART.

The font of the framerate indicator in the top right corner is a free to
redistribute font called "monospace," which was downloaded directly from
the Baking Pi[1] website. I also changed the color of the sky and the
walls in the scene to demonstrate the Raspberry Pi framebuffer's ability
to display RGBA colors.

Finally, the majority of the code in this project was rewritten in C
because using C++ would be overkill in such a constrained environment
and might introduce some overhead. However, I kept the game objects'
object-oriented and polymorphism design by imitating the C++ class
mechanism with pointers.

[1] https://www.cl.cam.ac.uk/projects/raspberrypi/tutorials/os/downloads.html